+++ source/javascripts/components/unified-editor/Triggers/components/TriggerConditions.tsx
@@ -1,6 +1,11 @@
 import { Box, Tag, Text, Tooltip, TypeIconName } from '@bitrise/bitkit';
 import { Fragment } from 'react/jsx-runtime';
-import { Condition, ConditionType, LegacyConditionType, TriggerType } from '../TriggersPage/TriggersPage.types';
+import {
+  Condition,
+  ConditionType,
+  LegacyConditionType,
+  TriggerType,
+} from '../../../../pages/TriggersPage/components/TriggersPage/TriggersPage.types';
 
 type TriggerConditionsProps = {
   conditions: Condition[];
