+++ source/javascripts/pages/TriggersPage/components/LegacyTriggers/TriggerCard.tsx
@@ -4,7 +4,7 @@ import { useSortable } from '@dnd-kit/sortable';
 import { CSS } from '@dnd-kit/utilities';
 import { segmentTrack } from '@/utils/segmentTracking';
 import { TriggerItem } from '../TriggersPage/TriggersPage.types';
-import TriggerConditions from '../TargetBasedTriggers/TriggerConditions';
+import TriggerConditions from '../../../../components/unified-editor/Triggers/components/TriggerConditions';
 
 interface TriggerCardProps extends CardProps {
   isOverlay?: boolean;
